% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLagLassoNetworkReconstruction.R
\name{timeLaggedOrderedLassoSemiSupervisedNetwork}
\alias{timeLaggedOrderedLassoSemiSupervisedNetwork}
\title{timeLaggedOrderedLassoSemiSupervisedNetwork}
\usage{
timeLaggedOrderedLassoSemiSupervisedNetwork(exprDataList, adjMatrix,
  output = "expr.", maxLag = 2, lambdaEdge = 1, lambdaNonEdge = 1,
  self = TRUE, method = "Solve.QP", strongly.ordered = FALSE,
  rescale = TRUE, rescaleSeparately = FALSE, maxiter = 500,
  inneriter = 100, iter.gg = 100, cores = 1)
}
\arguments{
\item{exprDataList}{list of expression datasets (timepoints x genes) for p genes}

\item{adjMatrix}{prior network adjacency matrix}

\item{output}{expression output type ('expr' or 'change expr.')}

\item{maxLag}{maximum predictor lag}

\item{lambdaEdge}{a scalar edge penalization parameter.}

\item{lambdaNonEdge}{a scalar non-edge penalization parameter.}

\item{self}{if \code{TRUE}, include loops in time-lagged regression models}

\item{method}{underlying ordered lasso optimization method ('Solve.QP' or 'GG')}

\item{strongly.ordered}{if \code{TRUE} (\code{FALSE}), use the strongly (weakly) ordered lasso}

\item{rescale}{if \code{TRUE}, rescale input exprssion data}

\item{rescaleSeparately}{if \code{TRUE}, rescale each dataset separately}

\item{maxiter}{maximum number of  time lagged ordered lasso iterations}

\item{inneriter}{maximum number ordered lasso iterations}

\item{iter.gg}{maximum number of generalized gradient iterations}

\item{cores}{number of parallel cores}
}
\value{
a predicted adjacency matrix
}
\description{
Predicts a posterior gene regulatory network from a prior network and expression data based on the time-lagged ordered lasso.
}
